[CHERI-{MIPS,RISC-V}] Allow inlining of copies without capabilities
Now that TargetLowering::getOptimalMemOpType() has access to a
PreserveCheriTags value, we can inline copies >= sizeof(cap) if they are
known not to require tag-preserving behaviour. This means that for a
struct foo {long a; long b; } we can now inline a copy-assignment
whereas previously we assumed that it might contain capabilities and
always called memcpy.